requirement.txt
streamlit
streamlit-chat
google-generativeai